backend/requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
authlib
requests
certifi
werkzeug
langchain
langchain-core
langchain-groq
groq
pydantic
gunicorn
PyPDF2
streamlit
